+++ src/pages/signup/SignupPage.tsx
@@ -1,17 +1,5 @@
-import { useState, useMemo } from 'react';
 import * as styles from './SignupPage.css';
-import {
-  VALIDATION_RULES,
-  isKoreanOnly,
-  isMinLength,
-  filterKorean,
-  isValidYearFormat,
-  isValidMonthFormat,
-  isValidDayFormat,
-  isMinimumAge,
-  isValidDate,
-} from './utils/validation';
-import type { GenderOption } from './types/formOptions';
+import useSignupForm from './hooks/useSignupForm';
 import TitleNavBar from '@/shared/components/navBar/TitleNavBar.tsx';
 import TextField from '@/shared/components/textField/TextField.tsx';
 import CtaButton from '@/shared/components/button/ctaButton/CtaButton.tsx';
@@ -20,90 +8,25 @@ import ShowErrorMessage from '@/shared/components/button/showErrorButton/ShowErr
 import { ERROR_MESSAGES } from '@/shared/constants/clientErrorMessage.ts';
 
 const SignupPage = () => {
-  const [name, setName] = useState('');
-  const [birthYear, setBirthYear] = useState('');
-  const [birthMonth, setBirthMonth] = useState('');
-  const [birthDay, setBirthDay] = useState('');
-  const [gender, setGender] = useState<GenderOption | null>(null);
-
-  const isNameValid =
-    isKoreanOnly(name) && isMinLength(name, VALIDATION_RULES.NAME_MIN_LENGTH);
-  const isNameFormatInvalid = name !== '' && !isKoreanOnly(name);
-  const isNameLengthInvalid =
-    name !== '' &&
-    isKoreanOnly(name) &&
-    !isMinLength(name, VALIDATION_RULES.NAME_MIN_LENGTH);
-
-  const handleNameChange = (input: string) => {
-    const filtered = filterKorean(input);
-    setName(filtered);
-  };
-
-  const yearNum = Number.parseInt(birthYear, 10);
-  const monthNum = Number.parseInt(birthMonth, 10);
-  const dayNum = Number.parseInt(birthDay, 10);
-
-  const yearFormatError = birthYear !== '' && !isValidYearFormat(birthYear);
-
-  const yearAgeError = useMemo(() => {
-    if (birthYear === '' || yearFormatError) return false;
-    return !isMinimumAge(yearNum, monthNum, dayNum, VALIDATION_RULES.MIN_AGE);
-  }, [birthYear, yearFormatError, yearNum, monthNum, dayNum]);
-
-  const handleNumericChange =
-    (setter: (val: string) => void) => (val: string) => {
-      const numeric = val.replace(/\D/g, '');
-      setter(numeric);
-    };
-
-  const monthFieldError = useMemo(() => {
-    if (birthMonth === '') return false;
-    return !isValidMonthFormat(birthMonth) || monthNum < 1 || monthNum > 12;
-  }, [birthMonth, monthNum]);
-
-  const dayFieldError = useMemo(() => {
-    if (birthDay === '') return false;
-    if (!isValidDayFormat(birthDay)) return true;
-    if (!isValidMonthFormat(birthMonth) || !isValidYearFormat(birthYear)) {
-      return dayNum < 1 || dayNum > 31;
-    }
-    return !isValidDate(yearNum, monthNum, dayNum);
-  }, [birthDay, birthMonth, birthYear, dayNum, monthNum, yearNum]);
-
-  const validationResult = useMemo(() => {
-    const allFieldsFilled =
-      name !== '' &&
-      birthYear !== '' &&
-      birthMonth !== '' &&
-      birthDay !== '' &&
-      gender !== null;
-
-    const noErrors =
-      isNameValid &&
-      !yearFormatError &&
-      !yearAgeError &&
-      !monthFieldError &&
-      !dayFieldError;
-
-    return {
-      allFieldsFilled,
-      noErrors,
-      isFormValid: allFieldsFilled && noErrors,
-    };
-  }, [
+  const {
     name,
     birthYear,
     birthMonth,
     birthDay,
     gender,
-    isNameValid,
+    handleNameChange,
+    handleBirthYearChange,
+    handleBirthMonthChange,
+    handleBirthDayChange,
+    setGender,
+    isNameFormatInvalid,
+    isNameLengthInvalid,
     yearFormatError,
     yearAgeError,
     monthFieldError,
     dayFieldError,
-  ]);
-
-  const { isFormValid } = validationResult;
+    isFormValid,
+  } = useSignupForm();
 
   return (
     <form>
@@ -112,6 +35,7 @@ const SignupPage = () => {
       <div className={styles.container}>
         <h1 className={styles.title}>추가 회원가입 정보</h1>
 
+        {/* 이름 입력 */}
         <div className={styles.fieldbox}>
           <h2 className={styles.fieldtitle}>이름</h2>
           <TextField
@@ -130,6 +54,7 @@ const SignupPage = () => {
           )}
         </div>
 
+        {/* 생년월일 입력 */}
         <div className={styles.fieldbox}>
           <h2 className={styles.fieldtitle}>생년월일</h2>
           <div className={styles.flexbox}>
@@ -138,7 +63,7 @@ const SignupPage = () => {
               placeholder="YYYY"
               maxLength={4}
               value={birthYear}
-              onChange={handleNumericChange(setBirthYear)}
+              onChange={handleBirthYearChange} // 이렇게!
               isError={birthYear !== '' && (yearFormatError || yearAgeError)}
               inputMode="numeric"
             />
@@ -147,7 +72,7 @@ const SignupPage = () => {
               placeholder="MM"
               maxLength={2}
               value={birthMonth}
-              onChange={handleNumericChange(setBirthMonth)}
+              onChange={handleBirthMonthChange}
               isError={birthMonth !== '' && monthFieldError}
               inputMode="numeric"
             />
@@ -156,7 +81,7 @@ const SignupPage = () => {
               placeholder="DD"
               maxLength={2}
               value={birthDay}
-              onChange={handleNumericChange(setBirthDay)}
+              onChange={handleBirthDayChange}
               isError={birthDay !== '' && dayFieldError}
               inputMode="numeric"
             />
@@ -172,6 +97,7 @@ const SignupPage = () => {
           })()}
         </div>
 
+        {/* 성별 선택 */}
         <div className={styles.fieldbox}>
           <h2 className={styles.fieldtitle}>성별</h2>
           <div className={styles.flexbox}>
